misc: move screen recorder into FFmpegRecordHelper
