--- Assets/Scripts/Scenes/View/ScreenRecorder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d963cc5e51aef7d499512be3d87ab297
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
